Create ETL with tested e t & l functions

+++ src/secs/utilities.py
@@ -2,14 +2,16 @@ from pathlib import Path
 from shutil import copyfile
 from typing import Dict
 
-import icontract
+import prefect
 from prefect import task
 
 
 @task
-@icontract.require(lambda dirpath: (dirpath / "master_template.xlsx").exists())
-def create_master_excel(dirpath: Path) -> None:
+def create_master_excel(template_filepath: Path, master_filepath: Path) -> None:
 
-    template_master_excel = dirpath / "master_template.xlsx"
-    master_excel = dirpath / "master.xlsx"
-    copyfile(template_master_excel, master_excel)
+    logger = prefect.context.get("logger")
+    if master_filepath.exists():
+        logger.info(f"Skip copying as {master_filepath} already exists ...")
+    else:
+        logger.info(f"Copying {template_filepath} to {master_filepath} ...")
+        copyfile(template_master_excel, master_excel)
